requirements.txt
fabric3 < 1.12
wrapt < 1.11
